requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
joblib==0.14.1
numpy==1.18.2
oauthlib==3.1.0
PyKrige==1.5.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
urllib3==1.25.8
pandas
django-pandas
gunicorn==19.6.0
dj-database-url 
psycopg2
Pillow